chore(directory): refine file settings
